Remove support of neutron-lbaas from lb_loadbalancer (#1697)
Remove support of neutron-lbaas from lb_loadbalancer and
revert to old naming scheme for all lb_shared functions
