+++ openstack/resource_openstack_lb_loadbalancer_v2.go
@@ -9,8 +9,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 
-	octavialoadbalancers "github.com/gophercloud/gophercloud/openstack/loadbalancer/v2/loadbalancers"
-	neutronloadbalancers "github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/lbaas_v2/loadbalancers"
+	"github.com/gophercloud/gophercloud/openstack/loadbalancer/v2/loadbalancers"
 )
 
 func resourceLoadBalancerV2() *schema.Resource {
@@ -29,7 +28,6 @@ func resourceLoadBalancerV2() *schema.Resource {
 			Delete: schema.DefaultTimeout(5 * time.Minute),
 		},
 
-		DeprecationMessage: "Support for neutron-lbaas will be removed. Make sure to use Octavia",
 		Schema: map[string]*schema.Schema{
 			"region": {
 				Type:     schema.TypeString,
@@ -128,9 +126,9 @@ func resourceLoadBalancerV2() *schema.Resource {
 
 func resourceLoadBalancerV2Create(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	config := meta.(*Config)
-	lbClient, err := chooseLBV2Client(d, config)
+	lbClient, err := config.LoadBalancerV2Client(GetRegion(d, config))
 	if err != nil {
-		return diag.Errorf("Error creating OpenStack networking client: %s", err)
+		return diag.Errorf("Error creating OpenStack loadbalancing client: %s", err)
 	}
 
 	var (
@@ -145,58 +143,37 @@ func resourceLoadBalancerV2Create(ctx context.Context, d *schema.ResourceData, m
 
 	adminStateUp := d.Get("admin_state_up").(bool)
 
-	if lbClient.Type == octaviaLBClientType {
-		createOpts := octavialoadbalancers.CreateOpts{
-			Name:         d.Get("name").(string),
-			Description:  d.Get("description").(string),
-			VipNetworkID: d.Get("vip_network_id").(string),
-			VipSubnetID:  d.Get("vip_subnet_id").(string),
-			VipPortID:    d.Get("vip_port_id").(string),
-			ProjectID:    d.Get("tenant_id").(string),
-			VipAddress:   d.Get("vip_address").(string),
-			AdminStateUp: &adminStateUp,
-			FlavorID:     d.Get("flavor_id").(string),
-			Provider:     lbProvider,
-		}
-
-		// availability_zone requires octavia minor version 2.14. Only set when specified.
-		if v, ok := d.GetOk("availability_zone"); ok {
-			aZ := v.(string)
-			createOpts.AvailabilityZone = aZ
-		}
-
-		if v, ok := d.GetOk("tags"); ok {
-			tags := v.(*schema.Set).List()
-			createOpts.Tags = expandToStringSlice(tags)
-		}
-
-		log.Printf("[DEBUG][Octavia] openstack_lb_loadbalancer_v2 create options: %#v", createOpts)
-		lb, err := octavialoadbalancers.Create(lbClient, createOpts).Extract()
-		if err != nil {
-			return diag.Errorf("Error creating openstack_lb_loadbalancer_v2: %s", err)
-		}
-		lbID = lb.ID
-		vipPortID = lb.VipPortID
-	} else {
-		createOpts := neutronloadbalancers.CreateOpts{
-			Name:         d.Get("name").(string),
-			Description:  d.Get("description").(string),
-			VipSubnetID:  d.Get("vip_subnet_id").(string),
-			TenantID:     d.Get("tenant_id").(string),
-			VipAddress:   d.Get("vip_address").(string),
-			AdminStateUp: &adminStateUp,
-			FlavorID:     d.Get("flavor_id").(string),
-			Provider:     lbProvider,
-		}
-
-		log.Printf("[DEBUG][Neutron] openstack_lb_loadbalancer_v2 create options: %#v", createOpts)
-		lb, err := neutronloadbalancers.Create(lbClient, createOpts).Extract()
-		if err != nil {
-			return diag.Errorf("Error creating openstack_lb_loadbalancer_v2: %s", err)
-		}
-		lbID = lb.ID
-		vipPortID = lb.VipPortID
+	createOpts := loadbalancers.CreateOpts{
+		Name:         d.Get("name").(string),
+		Description:  d.Get("description").(string),
+		VipNetworkID: d.Get("vip_network_id").(string),
+		VipSubnetID:  d.Get("vip_subnet_id").(string),
+		VipPortID:    d.Get("vip_port_id").(string),
+		ProjectID:    d.Get("tenant_id").(string),
+		VipAddress:   d.Get("vip_address").(string),
+		AdminStateUp: &adminStateUp,
+		FlavorID:     d.Get("flavor_id").(string),
+		Provider:     lbProvider,
 	}
+
+	// availability_zone requires octavia minor version 2.14. Only set when specified.
+	if v, ok := d.GetOk("availability_zone"); ok {
+		aZ := v.(string)
+		createOpts.AvailabilityZone = aZ
+	}
+
+	if v, ok := d.GetOk("tags"); ok {
+		tags := v.(*schema.Set).List()
+		createOpts.Tags = expandToStringSlice(tags)
+	}
+
+	log.Printf("[DEBUG] openstack_lb_loadbalancer_v2 create options: %#v", createOpts)
+	lb, err := loadbalancers.Create(lbClient, createOpts).Extract()
+	if err != nil {
+		return diag.Errorf("Error creating openstack_lb_loadbalancer_v2: %s", err)
+	}
+	lbID = lb.ID
+	vipPortID = lb.VipPortID
 
 	// Wait for load-balancer to become active before continuing.
 	timeout := d.Timeout(schema.TimeoutCreate)
@@ -222,55 +199,35 @@ func resourceLoadBalancerV2Create(ctx context.Context, d *schema.ResourceData, m
 
 func resourceLoadBalancerV2Read(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	config := meta.(*Config)
-	lbClient, err := chooseLBV2Client(d, config)
+	lbClient, err := config.LoadBalancerV2Client(GetRegion(d, config))
 	if err != nil {
-		return diag.Errorf("Error creating OpenStack networking client: %s", err)
+		return diag.Errorf("Error creating OpenStack loadbalancing client: %s", err)
 	}
 
 	var vipPortID string
 
-	if lbClient.Type == octaviaLBClientType {
-		lb, err := octavialoadbalancers.Get(lbClient, d.Id()).Extract()
-		if err != nil {
-			return diag.FromErr(CheckDeleted(d, err, "Unable to retrieve openstack_lb_loadbalancer_v2"))
-		}
-
-		log.Printf("[DEBUG][Octavia] Retrieved openstack_lb_loadbalancer_v2 %s: %#v", d.Id(), lb)
-
-		d.Set("name", lb.Name)
-		d.Set("description", lb.Description)
-		d.Set("vip_subnet_id", lb.VipSubnetID)
-		d.Set("vip_network_id", lb.VipNetworkID)
-		d.Set("tenant_id", lb.ProjectID)
-		d.Set("vip_address", lb.VipAddress)
-		d.Set("vip_port_id", lb.VipPortID)
-		d.Set("admin_state_up", lb.AdminStateUp)
-		d.Set("flavor_id", lb.FlavorID)
-		d.Set("loadbalancer_provider", lb.Provider)
-		d.Set("availability_zone", lb.AvailabilityZone)
-		d.Set("region", GetRegion(d, config))
-		d.Set("tags", lb.Tags)
-		vipPortID = lb.VipPortID
-	} else {
-		lb, err := neutronloadbalancers.Get(lbClient, d.Id()).Extract()
-		if err != nil {
-			return diag.FromErr(CheckDeleted(d, err, "Unable to retrieve openstack_lb_loadbalancer_v2"))
-		}
-
-		log.Printf("[DEBUG][Neutron] Retrieved openstack_lb_loadbalancer_v2 %s: %#v", d.Id(), lb)
-
-		d.Set("name", lb.Name)
-		d.Set("description", lb.Description)
-		d.Set("vip_subnet_id", lb.VipSubnetID)
-		d.Set("tenant_id", lb.TenantID)
-		d.Set("vip_address", lb.VipAddress)
-		d.Set("vip_port_id", lb.VipPortID)
-		d.Set("admin_state_up", lb.AdminStateUp)
-		d.Set("flavor_id", lb.FlavorID)
-		d.Set("loadbalancer_provider", lb.Provider)
-		d.Set("region", GetRegion(d, config))
-		vipPortID = lb.VipPortID
+	lb, err := loadbalancers.Get(lbClient, d.Id()).Extract()
+	if err != nil {
+		return diag.FromErr(CheckDeleted(d, err, "Unable to retrieve openstack_lb_loadbalancer_v2"))
 	}
+
+	log.Printf("[DEBUG] Retrieved openstack_lb_loadbalancer_v2 %s: %#v", d.Id(), lb)
+
+	d.Set("name", lb.Name)
+	d.Set("description", lb.Description)
+	d.Set("vip_subnet_id", lb.VipSubnetID)
+	d.Set("vip_network_id", lb.VipNetworkID)
+	d.Set("tenant_id", lb.ProjectID)
+	d.Set("vip_address", lb.VipAddress)
+	d.Set("vip_port_id", lb.VipPortID)
+	d.Set("admin_state_up", lb.AdminStateUp)
+	d.Set("flavor_id", lb.FlavorID)
+	d.Set("loadbalancer_provider", lb.Provider)
+	d.Set("availability_zone", lb.AvailabilityZone)
+	d.Set("region", GetRegion(d, config))
+	d.Set("tags", lb.Tags)
+
+	vipPortID = lb.VipPortID
 
 	// Get any security groups on the VIP Port.
 	if vipPortID != "" {
@@ -288,17 +245,42 @@ func resourceLoadBalancerV2Read(ctx context.Context, d *schema.ResourceData, met
 
 func resourceLoadBalancerV2Update(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	config := meta.(*Config)
-	lbClient, err := chooseLBV2Client(d, config)
+	lbClient, err := config.LoadBalancerV2Client(GetRegion(d, config))
+	var hasChange bool
 	if err != nil {
 		return diag.Errorf("Error creating OpenStack networking client: %s", err)
 	}
 
-	updateOpts, err := chooseLBV2LoadbalancerUpdateOpts(d, config)
-	if err != nil {
-		return diag.Errorf("Error building openstack_lb_loadbalancer_v2 update options: %s", err)
+	var updateOpts loadbalancers.UpdateOpts
+
+	if d.HasChange("name") {
+		hasChange = true
+		name := d.Get("name").(string)
+		updateOpts.Name = &name
+	}
+	if d.HasChange("description") {
+		hasChange = true
+		description := d.Get("description").(string)
+		updateOpts.Description = &description
+	}
+	if d.HasChange("admin_state_up") {
+		hasChange = true
+		asu := d.Get("admin_state_up").(bool)
+		updateOpts.AdminStateUp = &asu
 	}
 
-	if updateOpts != nil {
+	if d.HasChange("tags") {
+		hasChange = true
+		if v, ok := d.GetOk("tags"); ok {
+			tags := v.(*schema.Set).List()
+			tagsToUpdate := expandToStringSlice(tags)
+			updateOpts.Tags = &tagsToUpdate
+		} else {
+			updateOpts.Tags = &[]string{}
+		}
+	}
+
+	if hasChange {
 		// Wait for load-balancer to become active before continuing.
 		timeout := d.Timeout(schema.TimeoutUpdate)
 		err = waitForLBV2LoadBalancer(ctx, lbClient, d.Id(), "ACTIVE", getLbPendingStatuses(), timeout)
@@ -308,7 +290,7 @@ func resourceLoadBalancerV2Update(ctx context.Context, d *schema.ResourceData, m
 
 		log.Printf("[DEBUG] Updating openstack_lb_loadbalancer_v2 %s with options: %#v", d.Id(), updateOpts)
 		err = resource.Retry(timeout, func() *resource.RetryError {
-			_, err = neutronloadbalancers.Update(lbClient, d.Id(), updateOpts).Extract()
+			_, err = loadbalancers.Update(lbClient, d.Id(), updateOpts).Extract()
 			if err != nil {
 				return checkForRetryableError(err)
 			}
@@ -343,7 +325,7 @@ func resourceLoadBalancerV2Update(ctx context.Context, d *schema.ResourceData, m
 
 func resourceLoadBalancerV2Delete(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	config := meta.(*Config)
-	lbClient, err := chooseLBV2Client(d, config)
+	lbClient, err := config.LoadBalancerV2Client(GetRegion(d, config))
 	if err != nil {
 		return diag.Errorf("Error creating OpenStack networking client: %s", err)
 	}
@@ -351,7 +333,7 @@ func resourceLoadBalancerV2Delete(ctx context.Context, d *schema.ResourceData, m
 	log.Printf("[DEBUG] Deleting openstack_lb_loadbalancer_v2 %s", d.Id())
 	timeout := d.Timeout(schema.TimeoutDelete)
 	err = resource.Retry(timeout, func() *resource.RetryError {
-		err = neutronloadbalancers.Delete(lbClient, d.Id()).ExtractErr()
+		err = loadbalancers.Delete(lbClient, d.Id(), loadbalancers.DeleteOpts{}).ExtractErr()
 		if err != nil {
 			return checkForRetryableError(err)
 		}
